Create a namespace per integration test
